+++ MBAM/__init__.py
@@ -1,6 +1,5 @@
 from .geodesic import Geodesic
-from .utils import initial_velocity, jacobian_func, Avv_func
-from . import finite_difference
+from .utils import initial_velocity
 
 __all__ = ["Geodesic"]
 __version__ = "0.0.1"
