+++ MBAM/finite_difference.py
@@ -1,6 +1,8 @@
-"""Finite difference formulas for estimating directional derivatives of a
-function.
+"""Contains functions to compute the derivative of the model with finite
+difference formula.
 """
+
+import numpy as np
 
 
 # Forward Difference Formulas
@@ -88,3 +90,82 @@ def AuvCD4(f, x, u, v, h):
         + f(x - 2 * h * u + 2 * h * v)
         - f(x - 2 * h * u - 2 * h * v)
     ) / (48 * h ** 2)
+
+
+################################################################################
+# Wrapper functions to convert the finite difference formulae above to the
+# functions required by Geodesic.
+
+
+def jacobian_func(f, M, N, deriv_fn=FD, h=0.1):
+    """A wrapper function convert functions to compute directional first
+    derivative to the jacobian function needed in :class:`~MBAM.geodesic`.
+
+    Parameters
+    ----------
+    f: callable ``f(x)``
+        Function to evaluate the model.
+    M: int
+        Number of predictions the model makes.
+    N: int
+        Number of parameters the model takes.
+    deriv_fn: callable ``deriv_fn(f, x, v, h)`` (optional)
+        A function to compute the directional derivative of the model.
+    h: float (optional)
+        Finite difference step size.
+
+    Returns
+    -------
+    jacobian: callable ``jacobian(x)``
+        A function that evaluates the Jacobian of the model ``r`` at parameter
+        ``x``.
+
+    Notes
+    -----
+    If the predictions are weighted differently, the argument ``f`` should be
+    the residual function, assuming that the weights are encoded in the
+    residual.
+    """
+
+    def jacobian(x):
+        # This matrix will be used to specify the direction of the derivative
+        eye = np.diag(np.ones(N))
+        # Compute the jacobian matrix
+        jac = np.empty((M, N))
+        for ii, v in enumerate(eye):
+            jac[:, ii] = deriv_fn(f, x, v, h)
+        return jac
+
+    return jacobian
+
+
+def Avv_func(f, Avv_fn=AvvCD, h=0.1):
+    """A wrapper function convert functions to compute directional second
+    derivative to the Avv function needed in :class:`~MBAM.geodesic`.
+
+    Parameters
+    ----------
+    f: callable ``f(x)``
+        Function to evaluate the model.
+    Avv_fn: callable ``Avv_fn(f, x, v, h)`` (optional)
+        A function to compute the directional second derivative of the model.
+    h: float (optional)
+        Finite difference step size.
+
+    Returns
+    -------
+    Avv: callable ``Avv(x, v)``
+        A function that evaluates the directional second derivative (Avv) of
+        the model ``r`` at parameter ``x`` in the ``v`` direction.
+
+    Notes
+    -----
+    If the predictions are weighted differently, the argument ``f`` should be
+    the residual function, assuming that the weights are encoded in the
+    residual.
+    """
+
+    def Avv(x, v):
+        return Avv_fn(f, x, v, h)
+
+    return Avv
